Refactors updateTags() out into helper function
